+++ src/main/java/com/wellbeignatwork/backend/ServiceImp/IQuestionService.java
@@ -6,7 +6,7 @@ import com.wellbeignatwork.backend.model.Quiz;
 import java.util.List;
 
 public interface IQuestionService {
-	 public List<Question> getQuestions(int id);
+	 public List<Question> getQuestions(Long id);
 	 public void addQuestion(String question, String answer, Long marks, Quiz quiz, String optA, String optB, String optC, String optD);
 	 public void saveQuestions();
 }
